# swagger_server/models/request_post_technical_control.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.technical_data import TechnicalData  # noqa: F401,E501
from swagger_server import util


class RequestPostTechnicalControl(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, technical_data: TechnicalData=None, channel: str=None, external_transaction_id: str=None):  # noqa: E501
        """RequestPostTechnicalControl - a model defined in Swagger

        :param technical_data: The technical_data of this RequestPostTechnicalControl.  # noqa: E501
        :type technical_data: TechnicalData
        :param channel: The channel of this RequestPostTechnicalControl.  # noqa: E501
        :type channel: str
        :param external_transaction_id: The external_transaction_id of this RequestPostTechnicalControl.  # noqa: E501
        :type external_transaction_id: str
        """
        self.swagger_types = {
            'technical_data': TechnicalData,
            'channel': str,
            'external_transaction_id': str
        }

        self.attribute_map = {
            'technical_data': 'technical_data',
            'channel': 'channel',
            'external_transaction_id': 'externalTransactionId'
        }
        self._technical_data = technical_data
        self._channel = channel
        self._external_transaction_id = external_transaction_id

    @classmethod
    def from_dict(cls, dikt) -> 'RequestPostTechnicalControl':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The RequestPostTechnicalControl of this RequestPostTechnicalControl.  # noqa: E501
        :rtype: RequestPostTechnicalControl
        """
        return util.deserialize_model(dikt, cls)

    @property
    def technical_data(self) -> TechnicalData:
        """Gets the technical_data of this RequestPostTechnicalControl.


        :return: The technical_data of this RequestPostTechnicalControl.
        :rtype: TechnicalData
        """
        return self._technical_data

    @technical_data.setter
    def technical_data(self, technical_data: TechnicalData):
        """Sets the technical_data of this RequestPostTechnicalControl.


        :param technical_data: The technical_data of this RequestPostTechnicalControl.
        :type technical_data: TechnicalData
        """

        self._technical_data = technical_data

    @property
    def channel(self) -> str:
        """Gets the channel of this RequestPostTechnicalControl.


        :return: The channel of this RequestPostTechnicalControl.
        :rtype: str
        """
        return self._channel

    @channel.setter
    def channel(self, channel: str):
        """Sets the channel of this RequestPostTechnicalControl.


        :param channel: The channel of this RequestPostTechnicalControl.
        :type channel: str
        """

        self._channel = channel

    @property
    def external_transaction_id(self) -> str:
        """Gets the external_transaction_id of this RequestPostTechnicalControl.


        :return: The external_transaction_id of this RequestPostTechnicalControl.
        :rtype: str
        """
        return self._external_transaction_id

    @external_transaction_id.setter
    def external_transaction_id(self, external_transaction_id: str):
        """Sets the external_transaction_id of this RequestPostTechnicalControl.


        :param external_transaction_id: The external_transaction_id of this RequestPostTechnicalControl.
        :type external_transaction_id: str
        """

        self._external_transaction_id = external_transaction_id