# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: technical-control-ms
  description: Technical control microservice
  version: 1.0.0
servers:
- url: /rest/technical-control-api/v1.0/
paths:
  /post/technical-control:
    post:
      tags:
      - Technical
      summary: Guarda el control tecnico en la base de datos.
      description: Guardado de control tecnico de ingreso
      operationId: post_technical
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RequestPostTechnicalControl"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.technical_controller.TechnicalView
components:
  schemas:
    RequestPostTechnicalControl:
      properties:
        technical_data:
          $ref: "#/components/schemas/TechnicalData"
        channel:
          type: string
          example: Tech control
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
    TechnicalData:
      type: object
      properties:
        created_at:
          type: string
          example: 2022-02-03 13:58:00
        reasons:
          type: array
          items:
            type: integer
          example: [1,2,3]
        project:
          type: array
          items:
            type: integer
          example: [1,2,3]
        id_truck_license:
          type: integer
          example: 1
        initial_gasoline:
          type: string
          example: "3/4"
        final_gasoline:
          type: string
          example: "3/4"
          nullable: true
        initial_km:
          type: string
          example: "1"
        driver_companion:
          type: array
          items:
            type: integer
          example: [1,2,3]
        route_point:
          type: string
        observations:
          type: string
          nullable: true
        created_by:
          type: string
          example: "dmales"
        initial_images:
          maxItems: 10
          type: array
          items:
            type: string
            format: binary
        final_images:
          maxItems: 10
          type: array
          items:
            type: string
            format: binary
      example:
        created_at: 2022-02-03 13:58:00
        id_level_gasoline: 4
        reasons: [1,2,3]
        project: [1,2,3]
        id_truck_license: 1
        initial_km: 15555
        driver_companion: [1,2,3]
        route_point: Expalsa
        observations: Sin observacion
        initial_images: []
        final_images: []
        created_by: dmales
    GenericResponse:
      properties:
        data:
          type: object
          nullable: true
          example: Estructura generica.
        errorCode:
          type: integer
          example: 0
        message:
          type: string
          example: Mensaje generico.
        internalTransactionId:
          type: string
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
      example:
        errorCode: 0
        internalTransactionId: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId: fcea920f7412b5da7be0cf42b8c93759
        message: Mensaje generico.
        data: {}
    DriverCompanion:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DriverCompanionData"
      example:
        data:
        - id_reason: 1
          reason: Trabajo programado
        - id_reason: 2
          reason: Razon
    DriverCompanionData:
      type: object
      properties:
        id_driver_companion:
          type: integer
          example: 1
        name:
          type: string
          example: Daniel
      example:
        id_driver_companion: 1
        name: Daniel
    Reasons:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReasonsData"
      example:
        data:
        - id_reason: 1
          reason: Trabajo programado
        - id_reason: 2
          reason: Razon
    ReasonsData:
      type: object
      properties:
        id_reason:
          type: integer
          example: 1
        reason:
          type: string
          example: Trabajo programado
      example:
        id_reason: 1
        reason: Trabajo programado
    Project:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProjectsData"
      example:
        data:
        - id_project: 1
          name: Expalsa
        - id_project: 2
          name: Biomar
    ProjectsData:
      type: object
      properties:
        id_project:
          type: integer
          example: 1
        name:
          type: string
          example: Expalsa
      example:
        id_project: 1
        name: Trabajo programado
    ResponseError:
      type: object
      properties:
        errorCode:
          type: integer
        internalTransactionId:
          type: string
        externalTransactionId:
          type: string
        message:
          type: string
    RequestGeneric:
      properties:
        channel:
          type: string
          example: ZENTINEL
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c9375